Move tests that don't depend on timelock being in a particular configuration to the single node configuration.
