+++ timelock-server/src/suiteTest/java/com/palantir/atlasdb/timelock/MultiNodePaxosTimeLockServerIntegrationTest.java
@@ -18,13 +18,10 @@ package com.palantir.atlasdb.timelock;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -38,17 +35,12 @@ import com.google.common.collect.ImmutableSortedMap;
 import com.palantir.atlasdb.timelock.suite.PaxosSuite;
 import com.palantir.atlasdb.timelock.util.ExceptionMatchers;
 import com.palantir.atlasdb.timelock.util.ParameterInjector;
-import com.palantir.atlasdb.transaction.impl.TransactionConstants;
 import com.palantir.lock.LockDescriptor;
 import com.palantir.lock.LockMode;
 import com.palantir.lock.StringLockDescriptor;
 import com.palantir.lock.v2.LeaderTime;
 import com.palantir.lock.v2.LockRequest;
 import com.palantir.lock.v2.LockToken;
-import com.palantir.lock.v2.StartIdentifiedAtlasDbTransactionRequest;
-import com.palantir.lock.v2.StartIdentifiedAtlasDbTransactionResponse;
-import com.palantir.lock.v2.StartTransactionRequestV4;
-import com.palantir.lock.v2.StartTransactionResponseV4;
 import com.palantir.lock.v2.TimelockService;
 
 @RunWith(Parameterized.class)
@@ -228,122 +220,4 @@ public class MultiNodePaxosTimeLockServerIntegrationTest {
         assertThat(ts1).isLessThan(ts2);
     }
 
-    @Test
-    public void startIdentifiedAtlasDbTransactionGivesUsTimestampsInSequence() {
-        UUID requestorUuid = UUID.randomUUID();
-        StartIdentifiedAtlasDbTransactionResponse firstResponse = startIdentifiedAtlasDbTransaction(requestorUuid);
-        StartIdentifiedAtlasDbTransactionResponse secondResponse = startIdentifiedAtlasDbTransaction(requestorUuid);
-
-        // Note that we technically cannot guarantee an ordering between the fresh timestamp on response 1 and the
-        // immutable timestamp on response 2. Most of the time, we will have IT on response 2 = IT on response 1
-        // < FT on response 1, as the lock token on response 1 has not expired yet. However, if we sleep for long
-        // enough between the first and second call that the immutable timestamp lock expires, then
-        // IT on response 2 > FT on response 1.
-        assertThat(ImmutableList.of(
-                firstResponse.immutableTimestamp().getImmutableTimestamp(),
-                firstResponse.startTimestampAndPartition().timestamp(),
-                secondResponse.startTimestampAndPartition().timestamp())).isSorted();
-        assertThat(ImmutableList.of(
-                firstResponse.immutableTimestamp().getImmutableTimestamp(),
-                secondResponse.immutableTimestamp().getImmutableTimestamp(),
-                secondResponse.startTimestampAndPartition().timestamp())).isSorted();
-    }
-
-    @Test
-    public void startIdentifiedAtlasDbTransactionGivesUsStartTimestampsInTheSamePartition() {
-        UUID requestorUuid = UUID.randomUUID();
-        StartIdentifiedAtlasDbTransactionResponse firstResponse = startIdentifiedAtlasDbTransaction(requestorUuid);
-        StartIdentifiedAtlasDbTransactionResponse secondResponse = startIdentifiedAtlasDbTransaction(requestorUuid);
-
-        assertThat(firstResponse.startTimestampAndPartition().partition())
-                .isEqualTo(secondResponse.startTimestampAndPartition().partition());
-    }
-
-    @Test
-    public void temporalOrderingIsPreservedWhenMixingStandardTimestampAndIdentifiedTimestampRequests() {
-        UUID requestorUuid = UUID.randomUUID();
-        List<Long> temporalSequence = ImmutableList.of(
-                cluster.getFreshTimestamp(),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestorUuid),
-                cluster.getFreshTimestamp(),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestorUuid),
-                cluster.getFreshTimestamp());
-
-        assertThat(temporalSequence).isSorted();
-    }
-
-    @Test
-    public void distinctClientsStillShareTheSameSequenceOfTimestamps() {
-        UUID requestor1 = UUID.randomUUID();
-        UUID requestor2 = UUID.randomUUID();
-
-        List<Long> temporalSequence = ImmutableList.of(
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestor1),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestor1),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestor2),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestor2),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestor1),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestor2),
-                getStartTimestampFromIdentifiedAtlasDbTransaction(requestor1));
-
-        assertThat(temporalSequence).isSorted();
-    }
-
-    @Test
-    public void temporalOrderingIsPreservedForBatchedStartTransactionRequests() {
-        UUID requestor = UUID.randomUUID();
-        List<Long> allTimestamps = new ArrayList<>();
-
-        allTimestamps.addAll(getSortedBatchedStartTimestamps(requestor, 1));
-        allTimestamps.addAll(getSortedBatchedStartTimestamps(requestor, 4));
-        allTimestamps.addAll(getSortedBatchedStartTimestamps(requestor, 20));
-
-        assertThat(allTimestamps).isSorted();
-    }
-
-    @Test
-    public void temporalOrderingIsPreservedBetweenDifferentRequestorsForBatchedStartTransactionRequests() {
-        UUID requestor = UUID.randomUUID();
-        UUID requestor2 = UUID.randomUUID();
-        List<Long> allTimestamps = new ArrayList<>();
-
-        allTimestamps.addAll(getSortedBatchedStartTimestamps(requestor, 1));
-        allTimestamps.addAll(getSortedBatchedStartTimestamps(requestor2, 4));
-        allTimestamps.addAll(getSortedBatchedStartTimestamps(requestor, 20));
-        allTimestamps.addAll(getSortedBatchedStartTimestamps(requestor2, 15));
-
-        assertThat(allTimestamps).isSorted();
-    }
-
-    @Test
-    public void batchedTimestampsShouldBeSeparatedByModulus() {
-        UUID requestor = UUID.randomUUID();
-
-        List<Long> sortedTimestamps = getSortedBatchedStartTimestamps(requestor, 10);
-
-        Set<Long> differences = IntStream.range(0, sortedTimestamps.size() - 1)
-                .mapToObj(i -> sortedTimestamps.get(i + 1) - sortedTimestamps.get(i))
-                .collect(Collectors.toSet());
-
-        assertThat(differences).containsOnly((long) TransactionConstants.V2_TRANSACTION_NUM_PARTITIONS);
-    }
-
-    private List<Long> getSortedBatchedStartTimestamps(UUID requestorUuid, int numRequestedTimestamps) {
-        StartTransactionRequestV4 request = StartTransactionRequestV4.createForRequestor(
-                requestorUuid,
-                numRequestedTimestamps);
-        StartTransactionResponseV4 response = cluster.namespacedClient().startTransactions(request);
-        return response.timestamps().stream()
-                .boxed()
-                .collect(Collectors.toList());
-    }
-
-    private StartIdentifiedAtlasDbTransactionResponse startIdentifiedAtlasDbTransaction(UUID requestorUuid) {
-        return cluster.startIdentifiedAtlasDbTransaction(
-                StartIdentifiedAtlasDbTransactionRequest.createForRequestor(requestorUuid));
-    }
-
-    private long getStartTimestampFromIdentifiedAtlasDbTransaction(UUID requestorUuid) {
-        return startIdentifiedAtlasDbTransaction(requestorUuid).startTimestampAndPartition().timestamp();
-    }
 }
